requirements.txt
amqp==5.2.0
aniso8601==9.0.1
APScheduler==3.10.4
async-timeout==4.0.3
billiard==4.2.0
blinker==1.7.0
cachelib==0.9.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.6.1
email_validator==2.1.1
Flask==3.0.3
Flask-Caching==2.1.0
Flask-Login==0.6.3
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.4.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.7
importlib_resources==6.4.0
itsdangerous==2.1.2
Jinja2==3.1.3
kombu==5.3.7
MarkupSafe==2.1.5
passlib==1.7.4
prompt-toolkit==3.0.43
python-dateutil==2.9.0.post0
python-magic==0.4.27
pytz==2024.1
redis==5.0.3
six==1.16.0
SQLAlchemy==2.0.29
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
validators==0.28.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.2
WTForms==3.1.2
